fix: remove title in some views, improve comments #97
