+++ src/features/Search/components/VisualNovelListOptions/VisualNovelListOptions.tsx
@@ -2,7 +2,6 @@ import React, { memo, useCallback, VFC } from 'react';
 import { Box, IconButton } from '@chakra-ui/react';
 import { Icon } from '../../../../components/Icon/Icon';
 import { VisualNovelListVariant } from '../VisualNovelList/VisualNovelList';
-import { buttonOptions } from './data';
 
 interface Props {
 
@@ -12,6 +11,47 @@ interface Props {
   /** Variant change handler.  */
   readonly onVariantChange: (variant: VisualNovelListVariant) => void;
 }
+
+interface ButtonOption {
+
+  /**
+   * Icon name.
+   */
+  readonly icon: string;
+
+  /**
+   * List variant.
+   */
+  readonly variant: VisualNovelListVariant;
+
+  /**
+   * Icon button label.
+   */
+  readonly label: string;
+}
+
+export const buttonOptions: ButtonOption[] = [
+  {
+    icon: 'carbon:menu',
+    variant: 'table',
+    label: 'Table',
+  },
+  {
+    icon: 'carbon:grid',
+    variant: 'cards',
+    label: 'Cards',
+  },
+  {
+    icon: 'carbon:show-data-cards',
+    variant: 'wide-cards',
+    label: 'Wide cards',
+  },
+  {
+    icon: 'carbon:horizontal-view',
+    variant: 'extended-cards',
+    label: 'Extended Cards',
+  },
+];
 
 /** Visual novel list options. */
 const VisualNovelListOptionsComponent: VFC<Props> = ({ activeVariant, onVariantChange }) => {
